[demux] rtp ssrc defined as uint32_t

+++ src/demux.h
@@ -16,8 +16,8 @@ struct demux_ops_s
 	int (*run)(demux_ctx_t *ctx);
 	const char *(*mime)(demux_ctx_t *ctx, int index);
 	void (*eventlistener)(demux_ctx_t *ctx, event_listener_t listener, void *arg);
-	int (*attach)(demux_ctx_t *ctx, int index, decoder_t *decoder);
-	decoder_t *(*estream)(demux_ctx_t *ctx, int index);
+	int (*attach)(demux_ctx_t *ctx, long index, decoder_t *decoder);
+	decoder_t *(*estream)(demux_ctx_t *ctx, long index);
 	void (*destroy)(demux_ctx_t *ctx);
 };
